refactor: category 등록 및 읽는 함수 리팩터링
